Update dependencies and constants

+++ src/cache/championship.rs
@@ -129,7 +129,7 @@ impl ChampionshipCache {
 #[async_trait]
 impl EntityCache for ChampionshipCache {
     type Entity = Championship;
-    const EXPIRATION: usize = REDIS_CACHE_EXPIRATION;
+    const EXPIRATION: u64 = REDIS_CACHE_EXPIRATION;
 
     #[inline(always)]
     async fn get(&self, id: &i32) -> AppResult<Option<Self::Entity>> {
