+++ src/cache/mod.rs
@@ -32,7 +32,7 @@ impl RedisCache {
 #[async_trait]
 pub trait EntityCache {
     type Entity;
-    const EXPIRATION: usize = REDIS_CACHE_EXPIRATION;
+    const EXPIRATION: u64 = REDIS_CACHE_EXPIRATION;
 
     async fn get(&self, id: &i32) -> AppResult<Option<Self::Entity>>;
     async fn set(&self, entity: &Self::Entity) -> AppResult<()>;
