+++ src/config/constants.rs
@@ -6,15 +6,15 @@ pub const GOOGLE_TOKEN_URL: &str = "https://oauth2.googleapis.com/token";
 pub const GOOGLE_USER_INFO: &str = "https://www.googleapis.com/oauth2/v2/userinfo";
 
 // Tokens
-pub const GENERIC_TOKEN_EXPIRATION: usize = 15 * 60;
-pub const REFRESH_TOKEN_EXPIRATION: usize = 15 * 60 * 24 * 30;
+pub const GENERIC_TOKEN_EXPIRATION: u64 = 15 * 60;
+pub const REFRESH_TOKEN_EXPIRATION: u64 = 15 * 60 * 24 * 30;
 
 // Redis
 pub const REDIS_USER_PREFIX: &str = "user";
-pub const REDIS_CACHE_EXPIRATION: usize = 60 * 60 * 24;
+pub const REDIS_CACHE_EXPIRATION: u64 = 60 * 60 * 24;
 pub const REDIS_CHAMPIONSHIP_PREFIX: &str = "championship";
 pub const REDIS_F123_PREFIX: &str = "f123:championships";
-pub const REDIS_F123_PERSISTENCE: usize = 15 * 60;
+pub const REDIS_F123_PERSISTENCE: u64 = 15 * 60;
 
 // F123 Service
 // Socket
